requirements.txt
opencv-python
numpy
Pillow
ultralytics
gradio
requests
pydub
huggingface_hub
PyMuPDF
scipy